// src/app/features/mastertables/tblItemMaster/models/tblItemMaster.model.ts
export interface TblItemMaster {
	fldId: number;
	fldItemName: string;
	fldItemNameUrdu: string;
	fldItemNameTamil: string;
	fldItemType: string;
	fldCategoryName: string;
	fldBaseUnit: string;
	fldAllowFractionalQuantity: boolean;
	fldSubUnitLabel: string;
	fldAllowFlatRupeeValueBooking: boolean;
	fldMinOrderQuantityStep: number;
	fldItemImagePath: string;
	fldIsActive: boolean;
	fldCreatedBy: number;
	fldCreatedDt: Date;
	fldModifiedBy: number;
	fldModifiedDt: Date;
	fldDeletedBy: number;
	fldDeletedDt: Date;
}